Actually changed ROUGHSIZE to int. Added icemc and classdict.* to Makefile's clean and distclean commands. Removed ReadRoughnessData() in icemc.cc.

+++ Settings.cc
@@ -502,7 +502,7 @@ void Settings::ReadInputs(ifstream &inputsfile, ofstream &foutput, Anita* anita1
     Tools::GetNextNumberAsString(inputsfile,foutput,number);
     ROUGHNESS=(int)atoi(number.c_str()); 
     Tools::GetNextNumberAsString(inputsfile,foutput,number);
-    ROUGHSIZE=(double)atoi(number.c_str()); 
+    ROUGHSIZE=(int)atoi(number.c_str()); 
     Tools::GetNextNumberAsString(inputsfile,foutput,number);
     FIRN=atoi(number.c_str()); 
     if (FIRN==0)
